server/run.py
import uvicorn
import os
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

if __name__ == "__main__":
    # 获取配置
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", "8000"))
    reload = os.getenv("RELOAD", "True").lower() == "true"
    
    print(f"Starting Whisper Web API on {host}:{port}")
    uvicorn.run("app.main:app", host=host, port=port, reload=reload) 