server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
faster-whisper==0.9.0
websockets==11.0.3
pydantic==2.4.2
ffmpeg-python==0.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
torch==2.1.0 